requirements.txt
streamlit>=1.30.0
gpxpy>=1.6.0
pykml>=0.2.0
requests>=2.31.0
folium>=0.15.0
streamlit-folium>=0.15.0
pandas>=2.0.0
polyline>=2.0.0
python-dotenv>=1.0.0
